fix: Refactor types folder
